+++ src/Examples/WpfFdc3/WpfFdc3.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Fdc3.Json\MorganStanley.Fdc3.Json.csproj" />
+    <ProjectReference Include="..\..\Fdc3.NewtonsoftJson\MorganStanley.Fdc3.NewtonsoftJson.csproj" />
     <ProjectReference Include="..\..\Fdc3\MorganStanley.Fdc3.csproj" />
   </ItemGroup>
