+++ src/Fdc3.NewtonsoftJson/MorganStanley.Fdc3.NewtonsoftJson.csproj
@@ -1,0 +1,49 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <RootNamespace>MorganStanley.Fdc3.NewtonsoftJson</RootNamespace>
+        <AssemblyName>MorganStanley.Fdc3.NewtonsoftJson</AssemblyName>
+        <VersionPrefix>2.0.0</VersionPrefix>
+        <VersionSuffix>alpha.3</VersionSuffix>
+        <Product>.NET FDC3 Newtonsoft JSON</Product>
+        <SignAssembly>true</SignAssembly>
+        <AssemblyOriginatorKeyFile>..\keypair.snk</AssemblyOriginatorKeyFile>
+        <PackageReadmeFile>README.md</PackageReadmeFile>
+        <TargetFramework>netstandard2.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <LangVersion>8.0</LangVersion>
+        <RepositoryUrl>https://github.com/morganstanley/fdc3-dotnet</RepositoryUrl>
+        <Description>.NET Standard 2.0 FDC3 JSON helpers based on Newtonsoft.Json for use with MorganStanley.Fdc3.</Description>
+        <Tags>FDC3</Tags>
+    </PropertyGroup>
+
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+        <WarningLevel>9999</WarningLevel>
+    </PropertyGroup>
+
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+        <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
+        <WarningLevel>9999</WarningLevel>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <None Include="..\..\LICENSE">
+            <Pack>True</Pack>
+            <PackagePath></PackagePath>
+        </None>
+        <None Include="..\..\README.md">
+            <Pack>True</Pack>
+            <PackagePath>\</PackagePath>
+        </None>
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\Fdc3\MorganStanley.Fdc3.csproj" />
+    </ItemGroup>
+
+</Project>
