+++ src/Fdc3/MorganStanley.Fdc3.csproj
@@ -4,7 +4,7 @@
 		<RootNamespace>MorganStanley.Fdc3</RootNamespace>
 		<AssemblyName>MorganStanley.Fdc3</AssemblyName>
 		<VersionPrefix>2.0.0</VersionPrefix>
-        <VersionSuffix>alpha.2</VersionSuffix>
+        <VersionSuffix>alpha.3</VersionSuffix>
 		<Product>.NET FDC3</Product>
 		<SignAssembly>true</SignAssembly>
 		<AssemblyOriginatorKeyFile>..\keypair.snk</AssemblyOriginatorKeyFile>
