+++ src/Tests/MorganStanley.Fdc3.NewtonsoftJson.Tests/Context/ChartTests.cs
@@ -14,23 +14,22 @@
 
 using MorganStanley.Fdc3.Context;
 
-namespace MorganStanley.Fdc3.Json.Tests.Context;
+namespace MorganStanley.Fdc3.NewtonsoftJson.Tests.Context;
 
-public class PortfolioTests : ContextSchemaTest
+public class ChartTests : ContextSchemaTest
 {
-    public PortfolioTests()
-        : base("https://fdc3.finos.org/schemas/2.0/portfolio.schema.json")
+    public ChartTests()
+        : base("https://fdc3.finos.org/schemas/2.0/chart.schema.json")
     {
     }
 
     [Fact]
-    public async void Portfolio_SerializedJsonMatchesSchema()
+    public async void Chart_SerializedJsonMatchesSchema()
     {
-        Portfolio portfolio = new Portfolio(
-            new Position[] { new Position(0, new Instrument(new InstrumentID() { Ticker = "ticker" })) },
-            null,
-            "portfolio");
-
-        await this.ValidateSchema(portfolio);
+        Instrument instrument = new Instrument(new InstrumentID { Ticker = "TICKER" });
+        TimeRange timeRange = new TimeRange(DateTime.Now.ToString("o"), DateTime.Now.ToString("o"));
+        var otherConfig = new { A = "Foo", B = "Bar" };
+        Chart chart = new Chart(new Instrument[] { instrument }, timeRange, otherConfig, ChartStyle.Line, null, "chart");
+        await this.ValidateSchema(chart);
     }
 }
