+++ src/Tests/MorganStanley.Fdc3.NewtonsoftJson.Tests/Context/ChatInitSettingsTests.cs
@@ -14,7 +14,7 @@
 
 using MorganStanley.Fdc3.Context;
 
-namespace MorganStanley.Fdc3.Json.Tests.Context;
+namespace MorganStanley.Fdc3.NewtonsoftJson.Tests.Context;
 
 public class ChatInitSettingsTests : ContextSchemaTest
 {
