+++ src/Tests/MorganStanley.Fdc3.NewtonsoftJson.Tests/Context/Examples/email-contact.json
