+++ src/Tests/MorganStanley.Fdc3.NewtonsoftJson.Tests/MorganStanley.Fdc3.NewtonsoftJson.Tests.csproj
@@ -1,0 +1,41 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+
+        <IsPackable>false</IsPackable>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <None Remove="Context\Examples\email-contact.json" />
+        <None Remove="Context\Examples\email-contactList.json" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <EmbeddedResource Include="Context\Examples\email-contact.json" />
+        <EmbeddedResource Include="Context\Examples\email-contactList.json" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
+        <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+        <PackageReference Include="Newtonsoft.Json.Schema" Version="3.0.14" />
+        <PackageReference Include="xunit" Version="2.4.2" />
+        <PackageReference Include="xunit.runner.visualstudio" Version="2.4.5">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+        <PackageReference Include="coverlet.collector" Version="3.2.0">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\..\Fdc3.NewtonsoftJson\MorganStanley.Fdc3.NewtonsoftJson.csproj" />
+        <ProjectReference Include="..\..\Fdc3\MorganStanley.Fdc3.csproj" />
+    </ItemGroup>
+
+</Project>
