+++ src/Tests/MorganStanley.Fdc3.Tests/MorganStanley.Fdc3.Tests.csproj
@@ -22,7 +22,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Fdc3.Json\MorganStanley.Fdc3.Json.csproj" />
     <ProjectReference Include="..\..\Fdc3\MorganStanley.Fdc3.csproj" />
   </ItemGroup>
